reorganised typeDefs folder
