+++ src/models/index.ts
@@ -6,8 +6,8 @@ export { Grade } from "./Grade";
 export { Group } from "./Group";
 export { Render } from "./Render";
 export { Role } from "./Role";
-export { Specialty } from "./Specialty";
+export { speciality } from "./speciality";
 export { Task } from "./Task";
 export { TaskAssignation } from "./TaskAssignation";
-export { User} from "./User";
+export { User } from "./User";
 export { UserDetails } from "./UserDetails";
