+++ src/resolvers/index.ts
@@ -1,10 +1,8 @@
-import { merge } from 'lodash'
-import { userResolvers } from './userResolvers'
-import { taskResolvers } from './taskResolvers'
+import { merge } from "lodash";
 
-const resolvers = merge(
-  userResolvers,
-  taskResolvers,
-)
+import { userResolvers } from "./userResolvers";
+import { taskResolvers } from "./taskResolvers";
+
+const resolvers = merge(userResolvers, taskResolvers);
 
 export default resolvers;
