+++ src/server.ts
@@ -1,6 +1,6 @@
 import { ApolloServer } from "apollo-server-express";
 import resolvers from "./resolvers";
-import typeDefs from "./graphql/typedefs";
+import typeDefs from "./typeDefs";
 
 export async function getApolloServer(): Promise<ApolloServer> {
   const apolloServer = new ApolloServer({
