+++ src/typeDefs/index.ts
@@ -1,0 +1,39 @@
+import {
+  authentication,
+  comment,
+  classroom,
+  correction,
+  grade,
+  group,
+  user,
+  userDetails,
+  render,
+  role,
+  speciality,
+  task,
+  taskAssignation,
+} from "./types";
+import { mutation } from "./mutation";
+import { query } from "./query";
+import { shared } from "./shared";
+
+const typeDefs = [
+  query,
+  mutation,
+  shared,
+  authentication,
+  comment,
+  classroom,
+  correction,
+  grade,
+  group,
+  user,
+  userDetails,
+  render,
+  role,
+  speciality,
+  task,
+  taskAssignation,
+];
+
+export default typeDefs;
