+++ src/typeDefs/shared.ts
@@ -1,0 +1,5 @@
+import { gql } from "apollo-server-express";
+
+export const shared = gql`
+  scalar DateTime
+`;
