+++ src/typeDefs/types/authentication.ts
@@ -1,0 +1,10 @@
+import { gql } from "apollo-server-express";
+
+export const authentication = gql`
+  type Authentication {
+    _id: ID!
+    password: String!
+    salt: String!
+    user: ID!
+  }
+`;
