+++ src/typeDefs/types/group.ts
@@ -1,0 +1,9 @@
+import { gql } from "apollo-server-express";
+
+export const group = gql`
+  type Group {
+    _id: ID!
+    groupname: String!
+    users: [ID!]
+  }
+`;
