+++ src/typeDefs/types/speciality.ts
@@ -1,0 +1,8 @@
+import { gql } from "apollo-server-express";
+
+export const speciality = gql`
+  type speciality {
+    _id: ID!
+    speciality_name: String!
+  }
+`;
