+++ src/typeDefs/types/task.ts
@@ -1,0 +1,15 @@
+import { gql } from "apollo-server-express";
+
+export const task = gql`
+  type Task {
+    _id: ID!
+    taskname: String!
+    creation_date: DateTime!
+    users: [User]
+  }
+
+  input InputTask {
+    taskname: String
+    users: [ID]
+  }
+`;
