+++ src/typeDefs/types/userDetails.ts
@@ -1,0 +1,11 @@
+import { gql } from "apollo-server-express";
+
+export const userDetails = gql`
+  type User_details {
+    _id: ID!
+    firstname: String!
+    lastname: String!
+    email: String!
+    user: ID!
+  }
+`;
